Promote Numba RNN-T activation reads to FP32
Numba-CUDA cannot implicitly unify BF16 with the FP32 dynamic-programming
state, so BF16 activations make kernel type inference diverge and the launch
fails with RecursionError. Widen activation reads at the point of use in
logp() and logp_duration(), which keeps `acts` narrow in memory and does the
arithmetic in FP32.

Also allocate the CTA reduction scratch as the module FP32 dtype rather than
the activation dtype, so reductions no longer accumulate in the input
precision.

Standard RNN-T now runs under BF16. Multi-blank and TDT still fail earlier,
in the workspace allocation. FP32 results are unchanged.

+++ nemo/collections/asr/parts/numba/rnnt_loss/utils/cuda_utils/gpu_rnnt_kernel.py
@@ -28,6 +28,7 @@
 
 import math
 
+import numba
 import torch
 from numba import cuda
 
@@ -61,13 +62,13 @@ def logp(
         The sum of logprobs[mb, t, u, v] + denom[mb, t, u]
     """
     col = (mb * maxT + t) * maxU + u
-    return denom[col] + acts[col * alphabet_size + v]
+    return denom[col] + numba.float32(acts[col * alphabet_size + v])
 
 
 @cuda.jit(device=True, inline=True)
 def logp_duration(acts: torch.Tensor, maxT: int, maxU: int, num_durations: int, mb: int, t: int, u: int, v: int):
     col = (mb * maxT + t) * maxU + u
-    return acts[col * num_durations + v]
+    return numba.float32(acts[col * num_durations + v])
 
 
 @cuda.jit()
